Adding a ResourceVolume for spawning resources
Resources are spawned randomly throughout the map and are able to be picked up. The type of the resource spawned is also random
The total resource amount is a configurable option in the SpawningVolume's Unity Inspector

In the future I would like to make it so resources of higher value are spawned further away, but value is yet to be determined so I will leave this as a nice-to-have.