Adding Notification System
The notification system works so that a player can add a notification to the system and anything listening to that system can choose how to display those notifications.

In this case, I have implemented a listener on the In-Game UI that adds and removes notifications from the UI.

I have made it so every time a player picks up a resource, that player's Notification UI reflects that it was picked up.